refactor(dnd): split monolithic context.rs into focused sub-modules
Split the 3,412-line context.rs into a context/ directory module with
5 focused files, each with a single clear responsibility:

- mod.rs: DragContext struct, constructors, registration, queries, ARIA,
  state structs, free functions, and all tests
- pointer.rs: pointer drag lifecycle, traversal computation, hysteresis
- keyboard.rs: keyboard drag navigation, container switching, merge toggle
- auto_scroll.rs: scroll velocity constants and computation
- provider.rs: DragContextProvider component and props

No public API changes. All 272 tests pass. Fields changed from private
to pub(super) for submodule access.

Addresses: Architecture — DragContext File Size and Responsibility
